requirements.txt
# Local development dependencies — mirrors what is installed inside Docker images.
# These are NOT used by the Docker build (Dockerfile.airflow and Dockerfile.spark
# install packages directly). Install here for IDE support and local testing:
#   python3 -m venv .venv && source .venv/bin/activate && pip install -r requirements.txt

kafka-python==2.0.2
dbt-core==1.7.9
dbt-snowflake==1.7.4
apache-airflow-providers-snowflake==5.2.1
